Fix problem reading some FAT32 partitions (#605175)
This problem was introduced with GParted 0.4.7 when the "-a"
option was removed from the dosfsck command.  The dosfsck command
is used to read the number of used sectors.  Instead we will now
use the "-n" option which does not write any changes to the file
system, and allows non-interactive operation.