Prevent visual disk display area from disappearing
Added a display refresh to the event queue to address this
problem.

Previously, the display area of the visual disk seemed to
disappear when enough operations had been added to the queue so
that a scrollbar was required (about 4 operations).

If the three lines in HBoxOperations::load_operations method for
"//make scrollwindow focus on the last operation in the list"
were commented out then this also prevented the problem from
occurring.

Interestingly, if I changed the "set_cursor" method to
a "scroll_to_row" method, then this problem does not occur for an
additional 2 operations (about 6 operations total).

Ubuntu launchpad bug 583746 - Partition graph disappears after a
                              swap partition has been deleted
https://bugs.launchpad.net/ubuntu/+source/gparted/+bug/583746