Recognise lvm command immediately when rescanning for supported actions
Rescanning for supported actions in the File System Support dialog was
not detecting the removal and restoration of the lvm command.  GParted
was only updating supported actions after refreshing all devices.

Checking for the lvm command was tied to the refresh of the LVM2_PV_Info
cache, hence the behaviour.  Fix by always checking for the lvm command
when requested.

Also remove extra new line from end of an error message in
load_lvm2_pv_info_cache().