Use kdesudo in gparted.desktop when gksu is not installed (#687813)
Normally gksu is not installed by default on KDE systems.  When
compiling and installing gparted on a KDE system without gksu, the
desktop menu entry would try to run gparted as the user, and not with
root privileges.

To fix this, add kdesudo to the search list of root privilege
invocation programs.

Closes Bug #687813 - Use kdesudo on KDE, as gksu is not installed by
                     default