Use gksudo, if available, in gparted.desktop
If gksu was installed on a KDE system (tested with kubuntu 12.04 LTS),
then gksu would prompt for the password three times and always fail to
load gparted.  In a prior version of kubuntu 8.04 LTS, gksu would
properly invoke gparted from the "K --> System --> GParted" menu
entry.

Since gksudo also works on KDE (and kubuntu 12.04 LTS), add gksudo to
the root privilege invocation search list ahead of gksu.

Note that gksu has been kept for backwards compatibility with systems
that provide only gksu and not gksudo.

Related to Bug #687813 - Use kdesudo on KDE, as gksu is not
                         installed by default